Separate routes and logic in separate directories.
