+++ employee-vci/src/logic/vciConfigProvider.ts
