+++ participation-cert-vci/src/app.ts
@@ -4,8 +4,8 @@ import path, { dirname } from "path";
 import Koa from "koa";
 import serve from "koa-static";
 
-import adminRoutes from "./admin/routes.js";
-import vciRoutes from "./vci/routes.js";
+import adminRoutes from "./routes/admin/routes.js";
+import vciRoutes from "./routes/vci/routes.js";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename).split("/src")[0];
