+++ participation-cert-vci/src/logic/credentialsConfigProvider.ts
