+++ participation-cert-vci/src/logic/participationCertificate.ts
