+++ participation-cert-vci/src/logic/vciConfigProvider.ts
