+++ proxy-vci/src/logic/credentialsConfigProvider.ts
