+++ proxy-vci/src/logic/identityCredential.ts
