+++ proxy-vci/src/routes/admin/routes.ts
@@ -1,6 +1,6 @@
 import Router from "koa-router";
 
-import adminHandler from "../../../common/src/routes/admin/handler.js"
+import adminHandler from "../../../../common/src/routes/admin/handler.js"
 
 const init = () => {
   const router = new Router();
