+++ proxy-vci/src/routes/vci/routesHandler.ts
@@ -5,12 +5,12 @@ import fs from "fs/promises";
 import Koa from "koa";
 import { v4 as uuidv4 } from "uuid";
 
-import oauth2 from "../auth.js";
-import store from "../store.js";
-import authStore, { getAuthCode } from "../../../common/src/store/authStore.js";
-import { CredentialIssuer } from "../../../common/src/oid4vci/credentialEndpoint/CredentialIssuer.js";
-import { configure } from "./credentialsConfigProvider.js";
-import { readLocalMetadataResource } from "../../../common/src/utils/resourceUtils.js";
+import oauth2 from "../../auth.js";
+import store from "../../store.js";
+import authStore, { getAuthCode } from "../../../../common/src/store/authStore.js";
+import { CredentialIssuer } from "../../../../common/src/oid4vci/credentialEndpoint/CredentialIssuer.js";
+import { configure } from "../../logic/credentialsConfigProvider.js";
+import { readLocalMetadataResource } from "../../../../common/src/utils/resourceUtils.js";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename).split("/src")[0];
